+++ platform/nim_core_v2_macos/macos/Classes/common/services/FLTAIService.cpp
@@ -148,6 +148,8 @@ flutter::EncodableMap convertNIMAIUser(
 
   resultMap.insert(std::make_pair("modelType", object->modelType));
 
+  resultMap.insert(std::make_pair("aiModelType", object->aiModelType));
+
   flutter::EncodableMap modelConfig = convertAIModelConfig(object->modelConfig);
   resultMap.insert(std::make_pair("modelConfig", modelConfig));
   return resultMap;