+++ platform/nim_core_v2_platform_interface/lib/src/platform_interface/team/team.g.dart
@@ -101,7 +101,7 @@ const _$NIMTeamChatBannedModeEnumMap = {
   NIMTeamChatBannedMode.unknown: -1,
   NIMTeamChatBannedMode.chatBannedModeNone: 0,
   NIMTeamChatBannedMode.chatBannedModeBannedNormal: 1,
-  NIMTeamChatBannedMode.chatBannedModeBannedAll: 2,
+  NIMTeamChatBannedMode.chatBannedModeBannedAll: 3,
 };
 
 TeamLeftReuslt _$TeamLeftReusltFromJson(Map<String, dynamic> json) =>